+++ flutter_app_lib/logger_provider/lib/src/logger_object/_logger_object.dart
@@ -1,0 +1,2 @@
+export 'logger_config.dart';
+export 'logger_factory.dart';