+++ flutter_app_lib/logger_provider/lib/src/rotation_object/_rotation_object.dart
@@ -1,0 +1,2 @@
+export 'rotation_config.dart';
+export 'rotation_logger_factory.dart';